backend/requirements1.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
motor==3.3.2
pymongo==4.6.1
dnspython==2.5.0

# Caching
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.2

# HTTP Client with HTTP/2 Support
httpx[http2]==0.27.0
h2==4.1.0

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# File Handling & Utils
python-multipart==0.0.9
python-dotenv==1.0.0

# CORS
starlette==0.36.3

# Async Support
anyio==4.2.0

# SSL Support
pyOpenSSL==24.0.0
certifi==2024.2.2
urllib3==2.2.0

# AI Integration
#open-interpreter==0.4.3
#pip install open-interpreter==0.3.14 

pandas==2.2.3

# Sahdev dependencies


plotly==5.24.1

langchain

langchain-community

sentence_transformers==3.3.1

sqlparse==0.5.3
#psycopg2==2.9.10
psycopg2-binary

pgvector==0.3.6
pyodbc==5.2.0

#others
diskcache==5.6.3